pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Oleg Cherkashin <impuls_64@mail.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
